Lots of changes!
Many changes to reflect changes in Haferlib. The new Log system has
replaced old System.out calls. Changed TabState to have a "workbench"
that is basically an active GUIElement that is selected with the
toolbar. Moved tab management code to a TabContainer class. Added a
LogDisplay class that displays log output.